klusters: fix int/qsizetype mismatch in the recluster channel restriction
Build fix for patch 0245.

    nSelect = std::min(nSelect, allowedCols.size());

nSelect is int; allowedCols is a QSet and Qt6's size() returns qsizetype, i.e.
long long. std::min deduces ONE type from both arguments, so there is no candidate
and the file does not compile. Cast to int, which is the house style here already
(klusters.cpp:2100 does the same).

Fixed in the same pass: `const int nChAvail = cv.size();` on the channel-variance
path. That one compiles -- it is an implicit narrowing rather than a deduction
failure -- so it would have gone in silently. It is now an explicit static_cast.

int is the right target: allowedCols holds at most nChan * featPerCh columns and cv
at most nChan channels, so nothing here approaches 2^31.

WHY THE TEST MISSED IT, which is the part worth recording. The standalone model used
std::set<int> and wrote

    nSelect = std::min<int>(nSelect, (int)allowed.size());

with an explicit template argument and a cast. The shipped line had neither, and
std::set<int>::size() returns size_t rather than qsizetype anyway. So the model
never contained the expression that failed -- it tested the fixed code before the
code was fixed. Modelling Qt logic with STL containers can validate the ARITHMETIC,
which it did (every rule in 0245 still holds), but it cannot validate the TYPES, and
a type is what broke. Where a model stands in for a Qt container, its size() type has
to be stood in for too, or the model quietly diverges at exactly the line that
matters.

Validation (g++ -O2 -Wall -Wextra, standalone): re-modelled with a stand-in whose
size() returns qsizetype as Qt6 defines it, so the failing expression is present and
commented with the exact error it produces. The cast was checked to agree with a
wide comparison across available = {0,1,6,24,4096} and ask = {1,4,7,99} -- the fix
changes types, not behaviour, and 0245's rules are untouched.

Qt6 is NOT in this sandbox, which is why a deduction failure against a Qt typedef
reached you rather than me.

Applies with git am at HEAD (4cf41db).

+++ src/klusters/src/klusters_recluster.cpp
@@ -377,7 +377,8 @@ void KlustersApp::slotRecluster(){
                     // reaching outside the selection to make up the number would
                     // defeat the point of having made one.
                     if(!allowedCols.isEmpty())
-                        nSelect = std::min(nSelect, allowedCols.size());
+                        nSelect = std::min(nSelect,
+                                           static_cast<int>(allowedCols.size()));
 
                     double topVar   = iv.isEmpty() ? 0.0 : iv[0].second;
                     double minVar   = topVar * 0.05;
@@ -430,7 +431,7 @@ void KlustersApp::slotRecluster(){
                             // than the curator asked.  It still sets the timestamp
                             // column below.  The cap is the number of channels
                             // actually available after the selection, not nCh.
-                            const int nChAvail = cv.size();
+                            const int nChAvail = static_cast<int>(cv.size());
                             const int nChSel = qBound(1, autoSelectNFeatures,
                                                       std::max(1, nChAvail));
                             for(int c = 0; c < cv.size() && c < nChSel; ++c){
